requirements.txt
streamlit==1.47.0
pymongo==4.4.1
bcrypt==4.0.1
streamlit-cookies-manager==0.2.0
cryptography==41.0.2
pandas==2.0.3
numpy==1.24.3
altair==5.0.1
pillow==9.5.0
pytz==2023.3
streamlit-drawable-canvas==0.9.2  # Added this line for the drawing features
certifi==2023.7.22  # Add this for SSL certificate verification
dnspython==2.4.2    # Add this for MongoDB DNS resolution

# New OCR and Image Processing Dependencies
# SciPy: Pinning to a version compatible with numpy==1.24.3
# SciPy 1.9.3 is a good candidate that officially supports NumPy 1.24.x
scipy==1.9.3

# OpenCV: A recent 4.x headless version should be compatible with numpy 1.24.x
opencv-python-headless==4.8.0.74 # A widely used stable version

# EasyOCR and its PyTorch backend (CPU-only for Streamlit Cloud general deployment)
# EasyOCR might pull in its own torch, but explicitly pinning ensures control
torch==2.1.0 # This version of torch often works well with EasyOCR and older NumPy
torchaudio==2.1.0
torchvision==0.16.0
easyocr==1.7.1 # A recent stable version of EasyOCR

# Pytesseract
pytesseract==0.3.10